requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
PySide6-Essentials==6.3.2
requests==2.28.1
shiboken6==6.3.2
soupsieve==2.3.2.post1
urllib3==1.26.12
vk-api==11.9.9
